feat: se agrega refactorizacion

+++ ms-client/src/main/java/com/ntt/data/ms/client/service/serviceImpl/CustomerServiceImpl.java
@@ -4,12 +4,12 @@ import com.ntt.data.ms.client.client.BankAccountDTO;
 import com.ntt.data.ms.client.client.CreditDTO;
 import com.ntt.data.ms.client.config.CustomException;
 import com.ntt.data.ms.client.dto.BalanceAvailableDTO;
-import com.ntt.data.ms.client.dto.Movement;
-import com.ntt.data.ms.client.dto.ProductDTO;
 import com.ntt.data.ms.client.entity.Customer;
+import com.ntt.data.ms.client.mapper.ClientMapper;
+import com.ntt.data.ms.client.model.CustomerProductBalanceResponse;
+import com.ntt.data.ms.client.model.CustomerProductMovementsResponse;
 import com.ntt.data.ms.client.repository.CustomerRepository;
 import com.ntt.data.ms.client.service.CustomerService;
-import lombok.extern.java.Log;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
@@ -17,27 +17,27 @@ import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-import java.util.Collections;
-
 @Service
 @Slf4j
 public class CustomerServiceImpl implements CustomerService {
     private final CustomerRepository customerRepository;
     private final WebClient bankAccountApiClient;
     private final WebClient creditApiClient;
+    private final ClientMapper clientMapper;
 
     public CustomerServiceImpl(CustomerRepository customerRepository,
                                @Qualifier("bankAccountApiClient") WebClient bankAccountApiClient,
-                               @Qualifier("creditApiClient") WebClient creditApiClient) {
+                               @Qualifier("creditApiClient") WebClient creditApiClient,
+                               ClientMapper clientMapper) {
         this.customerRepository = customerRepository;
         this.bankAccountApiClient = bankAccountApiClient;
         this.creditApiClient = creditApiClient;
+        this.clientMapper = clientMapper;
     }
 
     @Override
-    public Mono<String> create(Customer customer) {
-        return customerRepository.save(customer)
-                .thenReturn("Cliente guardado con éxito");
+    public Mono<Customer> create(Customer customer) {
+        return customerRepository.save(customer);
     }
 
     @Override
@@ -71,10 +71,10 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
-    public Mono<String> update(Customer customer) {
+    public Mono<Customer> update(Customer customer) {
         // Asegurar que el _id sea un ObjectId
         if (customer.getId() == null) {
-            return Mono.error(new IllegalArgumentException("El ID no puede ser nulo"));
+            return Mono.error(new CustomException("El id no puede ser nulo"));
         }
 
         return customerRepository.findById(String.valueOf(customer.getId()))
@@ -82,18 +82,24 @@ public class CustomerServiceImpl implements CustomerService {
                     customer.setId(existingCliente.getId()); // Asegurar que se usa el mismo ID
                     return customerRepository.save(customer);
                 })
-                .thenReturn("Cliente Actualizado con éxito")
-                .switchIfEmpty(Mono.error(new RuntimeException("Cliente no encontrado")));
+                .switchIfEmpty(Mono.error(new CustomException("Cliente no encontrado")));
     }
 
     @Override
     public Mono<String> delete(String id) {
-        return customerRepository.deleteById(id)
-                .thenReturn("Cliente Eliminado con éxito");
+        return customerRepository.existsById(id)
+                .flatMap(exists -> {
+                    if (exists) {
+                        return customerRepository.deleteById(id)
+                                .thenReturn("Cliente eliminado con éxito");
+                    } else {
+                        return Mono.just("Cliente no encontrado");
+                    }
+                });
     }
 
     @Override
-    public Mono<ProductDTO> getMovement(String clientId, String productId) {
+    public Mono<CustomerProductMovementsResponse> getMovement(String clientId, String productId) {
         return customerRepository.findById(clientId)
                 .switchIfEmpty(Mono.error(new RuntimeException("Cliente no encontrado")))
                 .flatMap(customer -> getProductDTOMono(productId, customer));
@@ -101,11 +107,52 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
-    public Mono<BalanceAvailableDTO> getBalanceAvailable(String clientId, String productId) {
-        return null;
+    public Mono<CustomerProductBalanceResponse> getBalanceAvailable(String clientId, String productId) {
+        return customerRepository.findById(clientId)
+                .switchIfEmpty(Mono.error(new RuntimeException("Cliente no encontrado")))
+                .flatMap(customer -> getCustomerProductBalanceResponse(productId, customer));
     }
 
-    private Mono<ProductDTO> getProductDTOMono(String productId, Customer customer) {
+    private Mono<CustomerProductBalanceResponse> getCustomerProductBalanceResponse(String productId, Customer customer) {
+        Mono<BankAccountDTO> bankAccountDTOMono = fetchBankAccount(productId)
+                .onErrorResume(throwable -> {
+                    log.error("Error al obtener saldo de cuenta bancaria: {}", throwable.getMessage());
+                    return Mono.empty();
+                });
+        Mono<CreditDTO> creditDTOMono = fetchCredit(productId)
+                .onErrorResume(throwable -> {
+                    log.error("Error al obtener saldo de cuenta de crédito: {}", throwable.getMessage());
+                    return Mono.empty();
+                });
+
+        return Mono.zip(
+                        bankAccountDTOMono.hasElement(),
+                        creditDTOMono.hasElement()
+                )
+                .flatMap(hasElements -> {
+                    boolean hasBankAccount = hasElements.getT1();
+                    boolean hasCreditAccount = hasElements.getT2();
+                    if (!hasBankAccount && !hasCreditAccount) {
+                        return Mono.error(new CustomException("No se pudo obtener el saldo"));
+                    }
+                    return Mono.zip(
+                                    bankAccountDTOMono.defaultIfEmpty(new BankAccountDTO()),
+                                    creditDTOMono.defaultIfEmpty(new CreditDTO())
+                            )
+                            .map(tuple -> {
+                                BankAccountDTO bankAccountDTO = tuple.getT1();
+                                CreditDTO creditDTO = tuple.getT2();
+                                CustomerProductBalanceResponse balanceResponse = new CustomerProductBalanceResponse();
+                                balanceResponse.name(customer.getName());
+                                balanceResponse.identification(customer.getIdentification());
+                                balanceResponse.typeCard(bankAccountDTO.getType() == null ? creditDTO.getType().name() : bankAccountDTO.getType().name());
+                                balanceResponse.availableBalance(bankAccountDTO.getType() == null ? creditDTO.getAvailableBalance() : bankAccountDTO.getBalance());
+                                return balanceResponse;
+                            });
+                });
+    }
+
+    private Mono<CustomerProductMovementsResponse> getProductDTOMono(String productId, Customer customer) {
         Mono<BankAccountDTO> bankAccountDTOMono = fetchBankAccount(productId)
                 .onErrorResume(throwable -> {
                     log.error("Error al obtener movimientos de cuenta bancaria: {}", throwable.getMessage());
@@ -134,12 +181,12 @@ public class CustomerServiceImpl implements CustomerService {
                             .map(tuple -> {
                                 BankAccountDTO bankAccountDTO = tuple.getT1();
                                 CreditDTO creditDTO = tuple.getT2();
-                                return new ProductDTO(
-                                        customer.getName(),
-                                        customer.getIdentification(),
-                                        bankAccountDTO.getMovements(),
-                                        creditDTO
-                                );
+                                CustomerProductMovementsResponse movementsResponse = new CustomerProductMovementsResponse();
+                                movementsResponse.setName(customer.getName());
+                                movementsResponse.setIdentification(customer.getIdentification());
+                                movementsResponse.setMovementsBankAccount(clientMapper.movementsBankAccounts(bankAccountDTO.getMovements()));
+                                movementsResponse.setMovementsCredit(clientMapper.movementsCredit(creditDTO));
+                                return movementsResponse;
                             });
                 });
     }
